+++ src/components/Home/charla.js
@@ -2,45 +2,13 @@
 const { createElement } = React;
 const html = htm.bind(createElement);
 //A: definimos las funciones de react como requiere este runtime
-
 import TarjetaTexto from "./tarjetaTexto.js";
-
-async function fetch_textosCharla(idCharla) {
-  //U: trae todos los textos de una charla
-  console.log("fetch_textosCharla", idCharla);
-  const res = await fetchConToken(
-    //A: funcion definida en auth-servicio para acceder a la API con un token
-    `https://si.podemosaprender.org/api/charla/${idCharla}`
-  );
-  return res;
-}
-
-function linkParaHashtag(hashtag) {
-  return `<a href="#" class="hashtagLink">${hashtag}</a>`;
-  //A: agrego la class "hashtagLink" para referenciarla en agregarOnClickHashtags()
-}
-
-function hashtagsALinks(textosCharla) {
-  //A: Recorro cada texto de la charla renderizada
-  //Y reemplazo cada '#tag' por el mismo '#tag' entre etiquetas <a></a> para poder clickearlos
-  textosCharla.forEach((textoItem) => {
-    textoItem.texto = textoItem.texto.replace(
-      /#([A-Za-z0-9_]+)/g,
-      linkParaHashtag
-    );
-  });
-  //DBG:console.info(textosCharla);
-  return textosCharla;
-}
-
-async function obtenerTextosDeCharla(charla) {
-  const res = await fetch_textosCharla(charla);
-  const data = await res.json();
-  //A: Paso como parámetro la lista de textos de la charla correspondiente a funcion hashtagsALinks()
-  const textosConHashtagsLinks = hashtagsALinks(data.textos);
-  //DBG: console.info("charla.js/obtenerTextosDeCharla ejecutado");
-  return textosConHashtagsLinks;
-}
+import {
+  fetch_textosCharla,
+  linkParaHashtag,
+  hashtagsALinks,
+  obtenerTextosDeCharla,
+} from "../../servicios/charlas.js";
 
 class Charla extends React.Component {
   state = { listaTextos: [] };
@@ -64,7 +32,7 @@ class Charla extends React.Component {
                 (txt) =>
                   html`<${TarjetaTexto}
                     key=${txt.pk}
-                    textoCharla=${txt}
+                    texto=${txt}
                     tarjetaTextoProps=${this.props.tarjetaTextoProps}
                   >
                   <//>`
